[fix] Update userId to long in SetChatAdministratorCustomTitle
Update the scaladoc in other places to reflect the previous changes.
Release 5.0.3

Closes #123

+++ core/src/com/bot4s/telegram/methods/SetChatAdministratorCustomTitle.scala
@@ -7,10 +7,10 @@ import com.bot4s.telegram.models.ChatId
  *
  * @param chat_id 	      Unique identifier for the target chat or username of the target supergroup (in the format @supergroupusername)
  * @param user_id 	      Unique identifier of the target user
- * @param custom_title 	New custom title for the administrator; 0-16 characters, emoji are not allowed
+ * @param custom_title 	  New custom title for the administrator; 0-16 characters, emoji are not allowed
  */
 case class SetChatAdministratorCustomTitle(
   chatId: ChatId,
-  userId: Int,
+  userId: Long,
   customTitle: String
 ) extends JsonRequest[Boolean]
